+++ src/main/java/com/example/demo/customer/repository/CustomerRepository.java
@@ -11,9 +11,9 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
 	@Query(value = "SELECT get_customer_balance(:customerId, NOW())", nativeQuery = true)
-	BigDecimal getCustomerBalance(@Param("customerId") Long customerId);
+	BigDecimal getCustomerBalance(@Param("customerId") Integer customerId);
 
     List<Customer> findByFirstNameContainingIgnoreCaseOrLastNameContainingIgnoreCase(String firstName, String lastName);
